# web/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MatchCreate": {
        "description": "Which posting the resume should be measured against.",
        "properties": {
          "document_id": {
            "format": "uuid",
            "title": "Document Id",
            "type": "string"
          }
        },
        "required": [
          "document_id"
        ],
        "title": "MatchCreate",
        "type": "object"
      },
      "MatchRead": {
        "description": "The result of one match.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "document_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Id"
          },
          "document_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Title"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "matched_evidence": {
            "additionalProperties": {
              "type": "string"
            },
            "default": {},
            "title": "Matched Evidence",
            "type": "object"
          },
          "matched_keywords": {
            "items": {
              "type": "string"
            },
            "title": "Matched Keywords",
            "type": "array"
          },
          "missing_keywords": {
            "items": {
              "type": "string"
            },
            "title": "Missing Keywords",
            "type": "array"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "title": "Notes",
            "type": "array"
          },
          "resume_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resume Id"
          },
          "retrieved_chunk_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Retrieved Chunk Ids",
            "type": "array"
          },
          "score": {
            "title": "Score",
            "type": "number"
          },
          "suggestions": {
            "items": {
              "type": "string"
            },
            "title": "Suggestions",
            "type": "array"
          }
        },
        "required": [
          "id",
          "document_id",
          "created_at",
          "score",
          "matched_keywords",
          "missing_keywords",
          "suggestions",
          "notes",
          "retrieved_chunk_ids"
        ],
        "title": "MatchRead",
        "type": "object"
      },
...
      "MatchSummary": {
        "description": "One row of the history: enough to choose which match to open.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "document_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Id"
          },
          "document_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Title"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "matched_count": {
            "title": "Matched Count",
            "type": "integer"
          },
          "missing_count": {
            "title": "Missing Count",
            "type": "integer"
          },
          "resume_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resume Id"
          },
          "score": {
            "title": "Score",
            "type": "number"
          }
        },
        "required": [
          "id",
          "document_id",
          "document_title",
          "resume_id",
          "score",
          "matched_count",
          "missing_count",
          "created_at"
        ],
        "title": "MatchSummary",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/matches": {
      "get": {
        "description": "List the caller's own matches, newest first, optionally for one posting.\n\nAn unknown or deleted posting gives an empty list, not 404: document_id narrows\nthe caller's own rows and is not a resource to look up.",
        "operationId": "list_matches_matches_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "document_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MatchSummary"
                  },
                  "title": "Response List Matches Matches Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Matches",
        "tags": [
          "matching"
        ]
      }
    },
    "/matches/{match_id}": {
      "get": {
        "description": "Return one stored match in full, or 404 if it is not the caller's.",
        "operationId": "read_match_matches__match_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "match_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Match Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Read Match",
        "tags": [
          "matching"
        ]
      }
    },
...
    "/resumes/{resume_id}/match": {
      "post": {
        "description": "Score one of the caller's resumes against a posting and suggest edits.",
        "operationId": "match_resumes__resume_id__match_post",
        "parameters": [
          {
            "in": "path",
            "name": "resume_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Resume Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MatchCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Match",
        "tags": [
          "matching"
        ]
      }
    },
...
  }
}